refactor(spo): fold SpoQuery into Query
Move all SPO query methods from separate SpoQuery struct into the main
Query struct, eliminating the need for MergedObject. This simplifies
the GraphQL schema structure as suggested in PR review.
